moving INavigationService
